+++ exporters/otlp/otlpgrpc/otlp_integration_test.go
@@ -337,7 +337,7 @@ func TestNewExporter_withMultipleAttributeTypes(t *testing.T) {
 		sdktrace.WithBatcher(
 			exp,
 			// add following two options to ensure flush
-			sdktrace.WithBatchTimeout(5),
+			sdktrace.WithBatchTimeout(5*time.Second),
 			sdktrace.WithMaxExportBatchSize(10),
 		),
 	)